Big Form Update
Got the branching form to work properly. There are two submit buttons, which properly sends the information to the proper pages. There are new pages, which are going to have information pertaining to the form, as they are mere extentions. The thanks file is simply to thank the user and to send them back to the home page after completing the form.